requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.6.2
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
numpy==2.0.0
openai==1.35.7
opencv-contrib-python==4.10.0.84
pillow==10.3.0
pydantic==2.7.4
pydantic_core==2.18.4
python-dotenv==1.0.1
six==1.16.0
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.2
